requirements.txt
discord.py
discord-py-slash-command
python-dotenv
ffmpeg-python
ffprobe-python
filetype
spotipy